+++ src/qianfan/model/model.py
@@ -31,7 +31,7 @@ from qianfan.resources import (
 )
 from qianfan.resources.console import consts as console_const
 from qianfan.resources.console.model import Model as ResourceModel
-from qianfan.runnable.base import ExecuteSerializable
+from qianfan.common.runnable.base import ExecuteSerializable
 from qianfan.utils import log_error, log_info, log_warn
