+++ src/qianfan/trainer/base.py
@@ -28,7 +28,7 @@ from typing import (
 )
 
 from qianfan.errors import InternalError, InvalidArgumentError
-from qianfan.runnable.base import ExecuteSerializable
+from qianfan.common.runnable.base import ExecuteSerializable
 from qianfan.trainer.consts import ActionState
 from qianfan.trainer.event import Event, EventHandler, dispatch_event
 from qianfan.utils import log_debug, log_error, utils
